signal-engine: fix trailing-SL establishment and other external-backtest gaps
Fixes found while confirming the CSV signal-backtest (POST /strategies/{id}/
backtest-signals) actually behaves correctly on real historical data:

- Trailing-stop block in _simulate_one_trade was gated on stop_loss_price
  already being set, so a trade that opened while its indicator (EMA/
  SuperTrend) sat on the wrong side of entry price - a real, common case
  the initial-stop guard correctly rejects - could NEVER get a stop-loss
  established later either, even once the indicator became valid days
  after entry. The position just ran unprotected to end_of_data. Verified
  against real TITAN/CHOLAFIN SuperTrend(10,1) data; CHOLAFIN's fixed exit
  lands exactly on the date independently observed on a live chart.
- SuperTrend's multiplier grid was typed list[int], rejecting any
  fractional value (e.g. 2.5) at the schema level before it could even
  reach the route. Widened to list[float]/float, with 'period' coerced
  back to a real int at the one place combos are built (list slicing
  needs a genuine int, not 20.0).
- 'daily' SL interval is now accepted for this backtest (which reads
  historical candles) even though the shared StopLossInterval type used
  by LIVE stop-loss checking deliberately excludes it (that path polls an
  intraday-only feed) - otherwise a swing/positional CSV could never be
  backtested against a daily-chart indicator at all.
- extractErrorDetail (api.ts) only handled FastAPI's string-shaped error
  detail; a raw Pydantic schema-validation 422 (list-shaped) fell through
  to a bare "HTTP 422" with zero information. Now formats both.
- Per-symbol skip reasons, per-symbol SL-candle fetch for stop_loss_
  method='indicator' (previously only fetched for 'previous_candle'),
  entry/exit timestamps formatted without seconds in the trades table.

+++ systems/signal-engine/backend/app/domain/generation/rule.py
@@ -729,8 +729,12 @@ class RuleBacktestGridRequest(BaseModel):
     # means stop_loss_indicator_params above applies as one fixed value to
     # every combo, same as before this field existed. Not merged onto
     # stop_loss_indicator_params - every key the sweep needs must be named
-    # here (see expand_stop_loss_grid's own docstring on why).
-    stop_loss_indicator_param_grid: Optional[dict[str, list[int]]] = None
+    # here (see expand_stop_loss_grid's own docstring on why). list[float]
+    # not list[int] - 'multiplier' (SuperTrend) is genuinely fractional
+    # (e.g. 2.5); 'period' is coerced back to a real int inside
+    # expand_stop_loss_grid before use (a container can't type its value
+    # list differently per dict key).
+    stop_loss_indicator_param_grid: Optional[dict[str, list[float]]] = None
     # A second sweep dimension for stop_loss_method='percent' instead -
     # candidate stop-loss percent values, e.g. [1, 1.5, 2, 2.5] to try 4
     # different SL percentages against every combination in param_grid
